Simplify lets_talk.sh config & docker volumes

+++ lets_talk.sh
@@ -7,16 +7,17 @@ usage: ./lets_talk.sh OR sudo PORt-80 ./lets_talk.sh
 Wrapper script for launching the "Let's Talk!" application with docker-compose.
 Probably cross-compatible with POSIX-compliant shells.
 
-optional arguments:
+actions:
   -h      show this help message and exit
-  -c      delete the MySQL database before starting, a new one will be created.
-          This flag doesn't work with Windows.
+  -d      execute "docker compose down --volumes" before starting, which resets the MySQL database.
   -b      (re-)build images.
+  -p      prune orphaned docker stuff.
   -n      don't start the server.
-  -p      prune orphaned docker stuff. (this is an alternative to "-c" on Windows.)
-  -d      start server with development configuration (mount server folders instead of copying).
-  -w      force use of Windows workarounds.
-  -u      ignore detection of Windows + Docker Toolbox, forces regular docker configuration.
+
+optional arguments:
+  -m      start server with development configuration (mount server folders instead of copying).
+  -t      mount /etc/localtime:ro in containers to correct timers. This is automatically set for Linux servers.
+  -w      prevent mounting of /etc/localtime.
   -i NIC  specify the network interface for private IPv4 address. (default: enp3s0)
           Examples: enp3s0, eth0, wlan1, wlp2s0
           Special case: if you are using docker-machine (which is employed by Docker Toolbox),
@@ -27,13 +28,13 @@ EOF
 [ "$1" = "--help" ] && help && exit 0
 
 # default values
-clean=0
-nic=enp3s0
-skip=0
-dev=0
+down=0
 build=0
 prune=0
-windows=0
+skip=0
+dev=0
+timez=0
+nic=enp3s0
 
 # output colors
 red=$(tput setaf 1)
@@ -42,36 +43,17 @@ reset=$(tput sgr0)
 bold=$(tput bold)
 underline=$(tput smul)
 
-function incompat () {
-  n=$skip
-  d=$dev
-  w=0 && [ "$windows" = "on" ] && w=1
-  u=0 && [ "$windows" = "off" ] && w=1
-  
-  for flag in "$@"; do 
-    eval value='$'$flag
-    if [ "$value" = "1" ]; then
-      help
-      echo "${red}Incompatible flags ${bold}-$opt${reset}${red} and ${bold}-$flag${reset}" >&2
-      exit 1
-    fi
-  done
-  unset value n d w u
-}
-
-while getopts ":hcbnpwudi:" opt; do
+while getopts ":hdbpnmtwi:" opt; do
   case $opt in
   h   ) help && exit 0           ;;
-  c   ) clean=1                  ;;
+  d   ) down=1                   ;;
   b   ) build=1                  ;;
-  n   ) skip=1                   ;;
   p   ) prune=1                  ;;
-  w   ) incompat c d
-    windows=on && nic=machine    ;;
-  u   ) incompat w
-    windows=off                  ;;
-  d   ) incompat n w
-    dev=1                        ;;
+  n   ) skip=1                   ;;
+
+  m   ) dev=1                    ;;
+  t   ) timez=1                  ;;
+  w   ) timez=no                 ;;
   i   ) nic=${OPTARG,,}          ;;
   \?  ) help && echo "${red}Invalid option: ${bold}-$OPTARG${reset}" >&2
     exit 1 ;;
@@ -98,30 +80,18 @@ elif [[ ${uname^^} = *"MINGW"* ]]; then # try to detect Docker Quickstart Termin
   has_machine=0 && which docker-machine 1> /dev/null 2> /dev/null && has_machine=1;
   if [ "$has_machine" = "1" ]; then 
     info="$info (Docker Quickstart terminal on Windows)"
-    if [ "$windows" = "off" ]; then
-      info="$warn detected environment to be $uname and docker-machine to be installed."
-      info="$info\nWindows workarounds have been disabled by the flag \"-u\", behavior might be unstable."
-    elif [ "$windows" = "0" ]; then
-      windows=1;
-      nic=machine;
-    fi  
+    nic=machine
   fi
   unset has_machine
-else
-  info="$warn found environment to be $uname but systemd was not detected."
-  info="$info\nThis setup has not been tested, behavior might be unstable."
 fi
 
-if [[ ${uname^^} = *"LINUX"* ]] && [ "$windows" = "on" ]; then
-  info="$warn found environment to be $uname but the flag \"-w\" forces the use of Windows workarounds."
+if [[ ${uname^^} = *"LINUX"* ]]; then
+  [ "$timez" != "no" ] && timez=1
 fi
+[ "$timez" = "no" ] && timez=0
 
 info="$info$reset"
 echo $info 1>&2
-
-if [ "$windows" = "off" ]; then windows=0;
-elif [ "$windows" = "on" ]; then windows=1;
-fi
 
 # IP addresses & URLs ===========================
 
@@ -170,26 +140,19 @@ fi
 
 set -e # quit on failure
 
+if [ "$down" = "1" ]; then
+  set -x
+  PORT=$PORT docker-compose down -v
+  { set +x; } 2> /dev/null
+fi
+
 if [ "$prune" = "1" ]; then
   echo "I hope you know what you're doing..."
   echo "If you use docker for other projects besides Let's Talk, \"-p\" is a danerous flag to set."
   set -x
-  PORT=$PORT docker-compose down -v
+  [ "$down" = 0 ] && PORT=$PORT docker-compose down -v
   docker system prune
   docker volume prune
-  { set +x; } 2> /dev/null
-fi
-
-# host mount point for MySQL data
-db=database/sql
-mkdir -p $db
-
-# purge database files if asked to
-if [ "$clean" = "1" ]; then
-  [ "$windows" = "1" ] && echo "$warn this action is useless when used with Windows workarounds."
-  set -x
-  sudo rm -r $db
-  mkdir $db
   { set +x; } 2> /dev/null
 fi
 
@@ -199,16 +162,15 @@ if [ "$build" = "1" ]; then
   { set +x; } 2> /dev/null
 fi
 
-if [ "$skip" = "0" ]; then
-  if [ "$windows" = "0" ]; then
-    if [ "$dev" = "1" ]; then
-      PORT=$PORT docker-compose --file dev-compose.yml up --abort-on-container-exit
-    else
-      PORT=$PORT docker-compose up --abort-on-container-exit # run normally
-    fi
-  else
-    PORT=$PORT docker-compose --file windows-compose.yml up --abort-on-container-exit
-  fi
-else
+if [ "$skip" = "1" ]; then
   echo 'The flag "-n" was specified. Exiting now...'
+  exit
+fi
+
+if [ "$dev" = "1" ]; then
+  PORT=$PORT docker-compose --file dev-compose.yml up --abort-on-container-exit
+elif [ "$timez" = "1" ]; then
+  PORT=$PORT docker-compose --file tz-compose.yml up --abort-on-container-exit
+else
+  PORT=$PORT docker-compose up --abort-on-container-exit # run normally
 fi
